It works! :smile:
Added Debucket Command so that the bucket's NBT can be dumped to the chat for debugging.

Requires fleshing out and doing checks on entities (for example, do not allow axolotls, fish, etc. to be bucketed by this plugin).

SpawnReason needs to be changed to be appropriate for the case, right now the original SpawnReason is kept.

Locale and Config required.
Permissions required.
Messages required.

Plan to eventually bundle this with a resource pack that includes mobs in buckets for custom item data.