WIP of making it pretty and organized.
Added a locale.yml that is loaded by a Locale class.
Config exists but not set up to load automagically yet.
Added spacing for BucketMobs.
Added Message that is assigned by the locale, started tryna use it.

Haven't tested but wanna see what GsonSerializer results in when you output NBT data to chat when using /debucket.